Small fixes and improvements via CodeQL scanning.

+++ elki-core-data/src/main/java/elki/data/projection/random/SimplifiedRandomHyperplaneProjectionFamily.java
@@ -149,7 +149,7 @@ public class SimplifiedRandomHyperplaneProjectionFamily implements RandomProject
      */
     private double[] projectDense(NumberVector in, double[] ret) {
       final int k = this.k;
-      final double dim = Math.min(buf.length, in.getDimensionality());
+      final int dim = Math.min(buf.length, in.getDimensionality());
       for(int i = 0; i < dim; i++) {
         final boolean[] row = mat[i];
         double vali = in.doubleValue(i);
